# remove bitvectors for BFS