```
# largest CC of a kronecker graph
julia> g = loadgraph("/tmp/testg.jlg")
{31206, 3679818} undirected simple Int64 graph

# current master (bitvector)
julia> @benchmark bfs_tree($g, 1)
BenchmarkTools.Trial:
  memory estimate:  13.39 MiB
  allocs estimate:  158712
  --------------
  minimum time:     45.268 ms (0.00% GC)
  median time:      53.741 ms (10.93% GC)
  mean time:        53.920 ms (7.43% GC)
  maximum time:     75.079 ms (8.45% GC)
  --------------
  samples:          93
  evals/sample:     1

# This PR - Vector{Bool}
julia> @benchmark bfs_tree($g, 1)
BenchmarkTools.Trial:
  memory estimate:  13.42 MiB
  allocs estimate:  158710
  --------------
  minimum time:     30.131 ms (0.00% GC)
  median time:      37.070 ms (14.10% GC)
  mean time:        36.964 ms (10.61% GC)
  maximum time:     47.310 ms (16.77% GC)
  --------------
  samples:          136
  evals/sample:     1
```