@sbromberger 2017-08-17
Turns out bitvectors are pretty slow compared to `Vector{Bool}`, with the advantage being moderate space saving. We're talking a temporary allocation difference of 7 bits per vertex, so I think we can handle that: a million vertices is < 1MByte.